@@ -14,15 +14,15 @@
         }
     ],
     "require": {
-        "php": ">=7",
-        "twig/twig": "~1.30",
-        "illuminate/support": "5.5.*|5.6.*|5.7.*|5.8.*",
-        "illuminate/view": "5.5.*|5.6.*|5.7.*|5.8.*"
+        "php": ">=7.1",
+        "twig/twig": "~2.0",
+        "illuminate/support": "^5.5|^6|^7",
+        "illuminate/view": "^5.5|^6|^7"
     },
     "require-dev": {
-        "laravel/framework": "5.0.*",
+        "laravel/framework": "5.5.*",
         "mockery/mockery": "0.9.*",
-        "phpunit/phpunit": "~4.0",
+        "phpunit/phpunit": "~6.0",
         "squizlabs/php_codesniffer": "~1.5",
         "satooshi/php-coveralls": "~0.6"
     },
@@ -38,7 +38,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "0.10-dev"
+            "dev-master": "0.11-dev"
         },
         "laravel": {
             "providers": [
