@@ -12,13 +12,13 @@
 namespace TwigBridge\Command;
 
 use Illuminate\Console\Command;
+use InvalidArgumentException;
+use RuntimeException;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Finder\Finder;
-use Twig_Error_Loader;
-use Twig_Error;
-use RuntimeException;
-use InvalidArgumentException;
+use Twig\Error\Error;
+use Twig\Error\LoaderError;
 
 /**
  * Artisan command to check the syntax of Twig templates.
@@ -92,7 +92,8 @@ class Lint extends Command
             while (!feof(STDIN)) {
                 $template .= fread(STDIN, 1024);
             }
-            if (ftell(STDIN) !== 0) {
+
+            if (!empty($template)) {
                 return $this->display([$this->validate($template)], $format);
             }
         }
@@ -103,7 +104,7 @@ class Lint extends Command
         foreach ($files as $file) {
             try {
                 $template = $this->getContents($file);
-            } catch (Twig_Error_Loader $e) {
+            } catch (LoaderError $e) {
                 throw new RuntimeException(sprintf('File or directory "%s" is not readable', $file));
             }
 
@@ -170,7 +171,7 @@ class Lint extends Command
      */
     protected function getContents($file)
     {
-        return $this->laravel['twig.loader']->getSource($file);
+        return $this->laravel['twig.loader']->getSourceContext($file);
     }
 
     /**
@@ -185,7 +186,7 @@ class Lint extends Command
     {
         try {
             $this->twig->parse($this->twig->tokenize($template, $file));
-        } catch (Twig_Error $e) {
+        } catch (Error $e) {
             return [
                 'template'  => $template,
                 'file'      => $file,
