@@ -11,15 +11,15 @@
 
 namespace TwigBridge\Extension\Laravel;
 
-use Twig_Extension;
-use Twig_SimpleFilter;
-use Twig_SimpleFunction;
-use Illuminate\Translation\Translator as LaravelTranslator;
+use Illuminate\Contracts\Translation\Translator as LaravelTranslator;
+use Twig\TwigFilter;
+use Twig\TwigFunction;
+use Twig\Extension\AbstractExtension;
 
 /**
  * Access Laravels translator class in your Twig templates.
  */
-class Translator extends Twig_Extension
+class Translator extends AbstractExtension
 {
     /**
      * @var \Illuminate\Translation\Translator
@@ -50,8 +50,8 @@ class Translator extends Twig_Extension
     public function getFunctions()
     {
        return [
-           new Twig_SimpleFunction('trans', [$this, 'trans'], ['is_safe' => ['html']]),
-           new Twig_SimpleFunction('trans_choice', [$this->translator, 'transChoice']),
+           new TwigFunction('trans', [$this, 'trans'], ['is_safe' => ['html']]),
+           new TwigFunction('trans_choice', [$this->translator, 'transChoice']),
        ];
     }
 
@@ -77,7 +77,7 @@ class Translator extends Twig_Extension
     public function getFilters()
     {
         return [
-            new Twig_SimpleFilter(
+            new TwigFilter(
                 'trans',
                 [$this->translator, 'trans'],
                 [
@@ -85,9 +85,9 @@ class Translator extends Twig_Extension
                     'is_safe'    => ['html'],
                 ]
             ),
-            new Twig_SimpleFilter(
+            new TwigFilter(
                 'trans_choice',
-                [$this->translator, 'transChoice'],
+                [$this->translator, 'choice'],
                 [
                     'pre_escape' => 'html',
                     'is_safe'    => ['html'],
