@@ -11,7 +11,7 @@
 
 namespace TwigBridge\Extension\Loader;
 
-use Twig_SimpleFunction;
+use Twig\TwigFunction;
 
 /**
  * Extension to expose defined functions to the Twig templates.
@@ -40,7 +40,7 @@ class Functions extends Loader
         foreach ($load as $method => $callable) {
             list($method, $callable, $options) = $this->parseCallable($method, $callable);
 
-            $function = new Twig_SimpleFunction(
+            $function = new TwigFunction(
                 $method,
                 function () use ($callable) {
                     return call_user_func_array($callable, func_get_args());
