@@ -13,7 +13,7 @@ namespace TwigBridge\Facade;
 use Illuminate\Support\Facades\Facade;
 
 /**
- * @see \Twig_Environment
+ * @see \Twig\Environment
  * @see \TwigBridge\Bridge
  */
 class Twig extends Facade
