@@ -12,8 +12,8 @@
 namespace TwigBridge\Command;
 
 use Illuminate\Console\Command;
+use Twig\Environment;
 use TwigBridge\Bridge;
-use Twig_Environment;
 
 /**
  * Artisan command to show details about the TwigBridge package.
@@ -41,7 +41,7 @@ class TwigBridge extends Command
      */
     public function handle()
     {
-        $this->line('<info>Twig</info> version        <comment>'.Twig_Environment::VERSION.'</comment>');
+        $this->line('<info>Twig</info> version        <comment>'.Environment::VERSION.'</comment>');
         $this->line('<info>Twig Bridge</info> version <comment>'.Bridge::BRIDGE_VERSION.'</comment>');
     }
 
