@@ -11,14 +11,14 @@
 
 namespace TwigBridge\Extension\Laravel;
 
-use Twig_Extension;
-use Twig_SimpleFunction;
 use Illuminate\Http\Request;
+use Twig\TwigFunction;
+use Twig\Extension\AbstractExtension;
 
 /**
  * Access Laravels input class in your Twig templates.
  */
-class Input extends Twig_Extension
+class Input extends AbstractExtension
 {
     /**
      * @var \Illuminate\Http\Request
@@ -49,10 +49,10 @@ class Input extends Twig_Extension
     public function getFunctions()
     {
         return [
-            new Twig_SimpleFunction('input_get', [$this->request, 'input']),
-            new Twig_SimpleFunction('input_old', [$this->request, 'old']),
-            new Twig_SimpleFunction('input_has', [$this->request, 'has']),
-            new Twig_SimpleFunction('old', [$this->request, 'old']),
+            new TwigFunction('input_get', [$this->request, 'input']),
+            new TwigFunction('input_old', [$this->request, 'old']),
+            new TwigFunction('input_has', [$this->request, 'has']),
+            new TwigFunction('old', [$this->request, 'old']),
         ];
     }
 }
